Added Home Page Etc
Created and added Home Page, along with nav bar and footer to flights and contact. Added basic layout elements to flights and contact. Styling done in cssCreated and added the Home Page, along with the navigation bar and footer to the Flights and Contact pages. Basic layout elements were added to Flights and Contact. Styling was done in CSS.